+++ tyrian_mobile/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -11,6 +11,7 @@ import device_info_plus
 import games_services
 import in_app_purchase_storekit
 import just_audio
+import multipeer_coop
 import screen_retriever_macos
 import shared_preferences_foundation
 import window_manager
@@ -22,6 +23,7 @@ func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   SwiftGamesServicesPlugin.register(with: registry.registrar(forPlugin: "SwiftGamesServicesPlugin"))
   InAppPurchasePlugin.register(with: registry.registrar(forPlugin: "InAppPurchasePlugin"))
   JustAudioPlugin.register(with: registry.registrar(forPlugin: "JustAudioPlugin"))
+  MultipeerCoopPlugin.register(with: registry.registrar(forPlugin: "MultipeerCoopPlugin"))
   ScreenRetrieverMacosPlugin.register(with: registry.registrar(forPlugin: "ScreenRetrieverMacosPlugin"))
   SharedPreferencesPlugin.register(with: registry.registrar(forPlugin: "SharedPreferencesPlugin"))
   WindowManagerPlugin.register(with: registry.registrar(forPlugin: "WindowManagerPlugin"))